server/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
Flask-Bcrypt==1.0.1
python-dotenv==1.1.1
Flask-SocketIO==5.5.1
psycopg2-binary==2.9.10
langchain==0.3.27
gunicorn
eventlet